Put renderer component into seperat file.
