+++ src/app/shared/cranix-shared.module.ts
@@ -14,6 +14,7 @@ import { PipesModule } from 'src/app/pipes/pipe-modules';
 import { AgGridModule } from 'ag-grid-angular';
 import { ActionBTNRenderer } from 'src/app/pipes/ag-action-renderer';
 import { ApplyBTNRenderer } from 'src/app/pipes/ag-apply-renderer';
+import { EventRenderer } from 'src/app/pipes/ag-event-renderer.ts';
 import { ApplyCheckBoxBTNRenderer } from 'src/app/pipes/ag-apply-checkbox-renderer';
 import { CustomerActionRenderer } from 'src/app/pipes/ag-customer-action-renderer';
 import { YesNoBTNRenderer } from 'src/app/pipes/ag-yesno-renderer';
@@ -80,6 +81,7 @@ import { simpleToolbarOptions } from 'src/app/shared/models/constants'
     DeviceActionBTNRenderer,
     DownloadSoftwaresComponent,
     EditBTNRenderer,
+    EventRenderer,
     FilesCollectComponent,
     FileSystemUsageRenderer,
     FilesUploadComponent,
